Rename to register-user.dto
